+++ backend/viewmodels/classroom.go
@@ -3,7 +3,7 @@ package viewmodels
 import "backend/utils"
 
 type ClassroomDto struct {
-	Id    uint8  `query:"-" json:"id,omitempty" xml:"id,omitempty" form:"id,omitempty"`
+	Id    uint   `query:"-" json:"id,omitempty" xml:"id,omitempty" form:"id,omitempty"`
 	Major string `query:"major,omitempty" json:"major,omitempty" xml:"major,omitempty" form:"major,omitempty"`
 	Level string `query:"level,omitempty" json:"level,omitempty" xml:"level,omitempty" form:"level,omitempty"`
 	Seq   uint8  `query:"seq,omitempty" json:"seq,omitempty" xml:"seq,omitempty" form:"seq,omitempty"`
